+++ android/Bottari/data/src/main/java/com/bottari/data/source/remote/TeamBottariItemsRemoteDataSource.kt
@@ -1,0 +1,68 @@
+package com.bottari.data.source.remote
+
+import com.bottari.data.model.remote.team.bottari.FetchTeamBottariStatusResponse
+import com.bottari.data.model.remote.team.bottari.item.request.AssignedItemsCreateRequest
+import com.bottari.data.model.remote.team.bottari.item.request.AssignedItemsUpdateRequest
+import com.bottari.data.model.remote.team.bottari.item.request.PersonalItemsCreateRequest
+import com.bottari.data.model.remote.team.bottari.item.request.SharedItemsCreateRequest
+import com.bottari.data.model.remote.team.bottari.item.request.TeamBottariItemCheckUpdateRequest
+import com.bottari.data.model.remote.team.bottari.item.request.TeamBottariItemDeleteRequest
+import com.bottari.data.model.remote.team.bottari.item.request.TeamBottariItemRemindRequest
+import com.bottari.data.model.remote.team.bottari.item.request.TeamBottariItemUnCheckUpdateRequest
+import com.bottari.data.model.remote.team.bottari.item.response.AssignedItemsFetchResponse
+import com.bottari.data.model.remote.team.bottari.item.response.PersonalItemsFetchResponse
+import com.bottari.data.model.remote.team.bottari.item.response.SharedItemsFetchResponse
+import com.bottari.data.model.remote.team.bottari.item.response.TeamBottariItemChecklistFetchResponse
+
+interface TeamBottariItemsRemoteDataSource {
+    suspend fun fetchTeamBottari(teamBottariId: Long): Result<TeamBottariItemChecklistFetchResponse>
+
+    suspend fun uncheckBottariItem(
+        bottariItemId: Long,
+        request: TeamBottariItemUnCheckUpdateRequest,
+    ): Result<Unit>
+
+    suspend fun checkBottariItem(
+        bottariItemId: Long,
+        request: TeamBottariItemCheckUpdateRequest,
+    ): Result<Unit>
+
+    suspend fun fetchTeamBottariStatus(id: Long): Result<FetchTeamBottariStatusResponse>
+
+    suspend fun sendRemindByItem(
+        id: Long,
+        type: TeamBottariItemRemindRequest,
+    ): Result<Unit>
+
+    suspend fun createTeamBottariSharedItem(
+        id: Long,
+        request: SharedItemsCreateRequest,
+    ): Result<Unit>
+
+    suspend fun createTeamBottariPersonalItem(
+        id: Long,
+        request: PersonalItemsCreateRequest,
+    ): Result<Unit>
+
+    suspend fun createTeamBottariAssignedItem(
+        id: Long,
+        request: AssignedItemsCreateRequest,
+    ): Result<Unit>
+
+    suspend fun deleteTeamBottariItem(
+        id: Long,
+        type: TeamBottariItemDeleteRequest,
+    ): Result<Unit>
+
+    suspend fun fetchTeamAssignedItems(teamBottariId: Long): Result<List<AssignedItemsFetchResponse>>
+
+    suspend fun fetchTeamSharedItems(teamBottariId: Long): Result<List<SharedItemsFetchResponse>>
+
+    suspend fun fetchTeamPersonalItems(teamBottariId: Long): Result<List<PersonalItemsFetchResponse>>
+
+    suspend fun saveTeamBottariAssignedItem(
+        teamBottariId: Long,
+        assignedItemId: Long,
+        request: AssignedItemsUpdateRequest,
+    ): Result<Unit>
+}
