+++ android/Bottari/di/src/main/java/com/bottari/di/usecase/TeamBottariItemsUseCaseProvider.kt
@@ -10,45 +10,49 @@ import com.bottari.domain.usecase.team.FetchTeamAssignedItemsUseCase
 import com.bottari.domain.usecase.team.FetchTeamChecklistUseCase
 import com.bottari.domain.usecase.team.FetchTeamPersonalItemsUseCase
 import com.bottari.domain.usecase.team.FetchTeamSharedItemsUseCase
+import com.bottari.domain.usecase.team.FetchTeamStatusUseCase
 import com.bottari.domain.usecase.team.SaveTeamBottariAssignedItemUseCase
 import com.bottari.domain.usecase.team.SendRemindByItemUseCase
 import com.bottari.domain.usecase.team.UncheckTeamBottariItemUseCase
 
-object TeamBottariItemUseCaseProvider {
+object TeamBottariItemsUseCaseProvider {
     val fetchTeamChecklistUseCase: FetchTeamChecklistUseCase by lazy {
-        FetchTeamChecklistUseCase(RepositoryProvider.teamBottariRepository)
+        FetchTeamChecklistUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val checkTeamBottariItemUseCase: CheckTeamBottariItemUseCase by lazy {
-        CheckTeamBottariItemUseCase(RepositoryProvider.teamBottariRepository)
+        CheckTeamBottariItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val uncheckTeamBottariItemUseCase: UncheckTeamBottariItemUseCase by lazy {
-        UncheckTeamBottariItemUseCase(RepositoryProvider.teamBottariRepository)
+        UncheckTeamBottariItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val sendRemindByItemUseCase: SendRemindByItemUseCase by lazy {
-        SendRemindByItemUseCase(RepositoryProvider.teamBottariRepository)
+        SendRemindByItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val fetchTeamPersonalItemsUseCase: FetchTeamPersonalItemsUseCase by lazy {
-        FetchTeamPersonalItemsUseCase(RepositoryProvider.teamBottariRepository)
+        FetchTeamPersonalItemsUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val fetchTeamAssignedItemsUseCase: FetchTeamAssignedItemsUseCase by lazy {
-        FetchTeamAssignedItemsUseCase(RepositoryProvider.teamBottariRepository)
+        FetchTeamAssignedItemsUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val fetchTeamSharedItemsUseCase: FetchTeamSharedItemsUseCase by lazy {
-        FetchTeamSharedItemsUseCase(RepositoryProvider.teamBottariRepository)
+        FetchTeamSharedItemsUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val createTeamSharedItemUseCase: CreateTeamSharedItemUseCase by lazy {
-        CreateTeamSharedItemUseCase(RepositoryProvider.teamBottariRepository)
+        CreateTeamSharedItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val createTeamPersonalItemUseCase: CreateTeamPersonalItemUseCase by lazy {
-        CreateTeamPersonalItemUseCase(RepositoryProvider.teamBottariRepository)
+        CreateTeamPersonalItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val createTeamAssignedItemUseCase: CreateTeamAssignedItemUseCase by lazy {
-        CreateTeamAssignedItemUseCase(RepositoryProvider.teamBottariRepository)
+        CreateTeamAssignedItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val deleteTeamBottariItemUseCase: DeleteTeamBottariItemUseCase by lazy {
-        DeleteTeamBottariItemUseCase(RepositoryProvider.teamBottariRepository)
+        DeleteTeamBottariItemUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
     val saveTeamBottariAssignedItemUseCase: SaveTeamBottariAssignedItemUseCase by lazy {
-        SaveTeamBottariAssignedItemUseCase(RepositoryProvider.teamBottariRepository)
+        SaveTeamBottariAssignedItemUseCase(RepositoryProvider.teamBottariItemsRepository)
+    }
+    val fetchTeamStatusUseCase: FetchTeamStatusUseCase by lazy {
+        FetchTeamStatusUseCase(RepositoryProvider.teamBottariItemsRepository)
     }
 }
