+++ android/Bottari/domain/src/main/java/com/bottari/domain/repository/TeamBottariRepository.kt
@@ -1,87 +1,14 @@
 package com.bottari.domain.repository
 
-import com.bottari.domain.model.bottari.item.BottariItem
 import com.bottari.domain.model.team.bottari.TeamBottari
-import com.bottari.domain.model.team.bottari.TeamBottariCheckList
 import com.bottari.domain.model.team.bottari.TeamBottariDetail
-import com.bottari.domain.model.team.bottari.TeamBottariStatus
-import com.bottari.domain.model.team.bottari.item.TeamBottariItemType
-import com.bottari.domain.model.team.member.TeamMember
-import com.bottari.domain.model.team.member.TeamMemberStatus
-import com.bottari.domain.model.team.member.TeamStatus
 
 interface TeamBottariRepository {
     suspend fun createTeamBottari(title: String): Result<Long?>
 
-    suspend fun fetchTeamBottari(teamBottariId: Long): Result<TeamBottariCheckList>
-
-    suspend fun uncheckBottariItem(
-        bottariItemId: Long,
-        type: String,
-    ): Result<Unit>
-
-    suspend fun checkBottariItem(
-        bottariItemId: Long,
-        type: String,
-    ): Result<Unit>
-
     suspend fun fetchTeamBottaries(): Result<List<TeamBottari>>
 
-    suspend fun fetchTeamMembers(id: Long): Result<TeamStatus>
-
     suspend fun fetchTeamBottariDetail(teamBottariId: Long): Result<TeamBottariDetail>
-
-    suspend fun fetchTeamBottariStatus(id: Long): Result<TeamBottariStatus>
-
-    suspend fun sendRemindByItem(
-        id: Long,
-        type: String,
-    ): Result<Unit>
-
-    suspend fun fetchTeamMembersStatus(id: Long): Result<List<TeamMemberStatus>>
-
-    suspend fun createTeamBottariSharedItem(
-        id: Long,
-        name: String,
-    ): Result<Unit>
-
-    suspend fun createTeamBottariPersonalItem(
-        id: Long,
-        name: String,
-    ): Result<Unit>
-
-    suspend fun createTeamBottariAssignedItem(
-        id: Long,
-        name: String,
-        teamMemberIds: List<Long>,
-    ): Result<Unit>
-
-    suspend fun deleteTeamBottariItem(
-        id: Long,
-        type: TeamBottariItemType,
-    ): Result<Unit>
-
-    suspend fun sendRemindByMemberMessage(
-        teamBottariId: Long,
-        memberId: Long,
-    ): Result<Unit>
-
-    suspend fun joinTeamBottari(inviteCode: String): Result<Unit>
-
-    suspend fun fetchTeamBottariMembers(teamBottariId: Long): Result<List<TeamMember>>
-
-    suspend fun fetchTeamAssignedItems(teamBottariId: Long): Result<List<BottariItem>>
-
-    suspend fun fetchTeamSharedItems(teamBottariId: Long): Result<List<BottariItem>>
-
-    suspend fun fetchTeamPersonalItems(teamBottariId: Long): Result<List<BottariItem>>
-
-    suspend fun saveTeamBottariAssignedItem(
-        teamBottariId: Long,
-        assignedItemId: Long,
-        name: String,
-        assigneeIds: List<Long>,
-    ): Result<Unit>
 
     suspend fun exitTeamBottari(teamBottariId: Long): Result<Unit>
 }
