+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/CheckTeamBottariItemUseCase.kt
@@ -1,12 +1,12 @@
 package com.bottari.domain.usecase.team
 
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class CheckTeamBottariItemUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
     suspend operator fun invoke(
         id: Long,
         type: String,
-    ): Result<Unit> = teamBottariRepository.checkBottariItem(id, type)
+    ): Result<Unit> = teamBottariItemsRepository.checkBottariItem(id, type)
 }
