+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/CreateTeamPersonalItemUseCase.kt
@@ -1,12 +1,12 @@
 package com.bottari.domain.usecase.team
 
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class CreateTeamPersonalItemUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
     suspend operator fun invoke(
         bottariId: Long,
         name: String,
-    ): Result<Unit> = teamBottariRepository.createTeamBottariPersonalItem(bottariId, name)
+    ): Result<Unit> = teamBottariItemsRepository.createTeamBottariPersonalItem(bottariId, name)
 }
