+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/CreateTeamSharedItemUseCase.kt
@@ -1,12 +1,12 @@
 package com.bottari.domain.usecase.team
 
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class CreateTeamSharedItemUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
     suspend operator fun invoke(
         bottariId: Long,
         name: String,
-    ): Result<Unit> = teamBottariRepository.createTeamBottariSharedItem(bottariId, name)
+    ): Result<Unit> = teamBottariItemsRepository.createTeamBottariSharedItem(bottariId, name)
 }
