+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/DeleteTeamBottariItemUseCase.kt
@@ -1,13 +1,13 @@
 package com.bottari.domain.usecase.team
 
 import com.bottari.domain.model.team.bottari.item.TeamBottariItemType
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class DeleteTeamBottariItemUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
     suspend operator fun invoke(
         itemId: Long,
         type: TeamBottariItemType,
-    ) = teamBottariRepository.deleteTeamBottariItem(itemId, type)
+    ) = teamBottariItemsRepository.deleteTeamBottariItem(itemId, type)
 }
