+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/FetchTeamChecklistUseCase.kt
@@ -1,10 +1,10 @@
 package com.bottari.domain.usecase.team
 
 import com.bottari.domain.model.team.bottari.TeamBottariCheckList
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class FetchTeamChecklistUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
-    suspend operator fun invoke(id: Long): Result<TeamBottariCheckList> = teamBottariRepository.fetchTeamBottari(id)
+    suspend operator fun invoke(id: Long): Result<TeamBottariCheckList> = teamBottariItemsRepository.fetchTeamBottari(id)
 }
