+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/FetchTeamPersonalItemsUseCase.kt
@@ -1,10 +1,10 @@
 package com.bottari.domain.usecase.team
 
 import com.bottari.domain.model.bottari.item.BottariItem
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class FetchTeamPersonalItemsUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
-    suspend operator fun invoke(bottariId: Long): Result<List<BottariItem>> = teamBottariRepository.fetchTeamPersonalItems(bottariId)
+    suspend operator fun invoke(bottariId: Long): Result<List<BottariItem>> = teamBottariItemsRepository.fetchTeamPersonalItems(bottariId)
 }
