+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/FetchTeamStatusUseCase.kt
@@ -1,10 +1,10 @@
 package com.bottari.domain.usecase.team
 
 import com.bottari.domain.model.team.bottari.TeamBottariStatus
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamBottariItemsRepository
 
 class FetchTeamStatusUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamBottariItemsRepository: TeamBottariItemsRepository,
 ) {
-    suspend operator fun invoke(id: Long): Result<TeamBottariStatus> = teamBottariRepository.fetchTeamBottariStatus(id)
+    suspend operator fun invoke(id: Long): Result<TeamBottariStatus> = teamBottariItemsRepository.fetchTeamBottariStatus(id)
 }
