+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/JoinTeamBottariUseCase.kt
@@ -1,9 +1,9 @@
 package com.bottari.domain.usecase.team
 
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamMemberRepository
 
 class JoinTeamBottariUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamMemberRepository: TeamMemberRepository,
 ) {
-    suspend operator fun invoke(inviteCode: String): Result<Unit> = teamBottariRepository.joinTeamBottari(inviteCode)
+    suspend operator fun invoke(inviteCode: String): Result<Unit> = teamMemberRepository.joinTeamBottari(inviteCode)
 }
