+++ android/Bottari/domain/src/main/java/com/bottari/domain/usecase/team/SendRemindByMemberMessageUseCase.kt
@@ -1,12 +1,12 @@
 package com.bottari.domain.usecase.team
 
-import com.bottari.domain.repository.TeamBottariRepository
+import com.bottari.domain.repository.TeamMemberRepository
 
 class SendRemindByMemberMessageUseCase(
-    private val teamBottariRepository: TeamBottariRepository,
+    private val teamMemberRepository: TeamMemberRepository,
 ) {
     suspend operator fun invoke(
         teamBottariId: Long,
         memberId: Long,
-    ): Result<Unit> = teamBottariRepository.sendRemindByMemberMessage(teamBottariId, memberId)
+    ): Result<Unit> = teamMemberRepository.sendRemindByMemberMessage(teamBottariId, memberId)
 }
